Removed config load from accessory (since we're using require() earlier in the file) Tweaked _deployment_hook_master to exit at the end. If debug is on OR user is super admin a message is displayed Added conditional checks against existing table before dbforge actions in setup model file

+++ deployment_hooks/models/deployment_hooks_setup_model.php
@@ -143,6 +143,11 @@ class Deployment_hooks_setup_model {
 	 */
 	public function create_dh_table()
 	{
+		if ($this->_EE->db->table_exists('deployment_hook_posts'))
+		{
+			return;
+		}
+		
 		$table_fields = array(
 			'deploy_id'        => array('type' => 'INT', 'constraint' => '10', 'unsigned' => TRUE, 'auto_increment' => TRUE),
 			'deploy_ip'        => array('type' => 'VARCHAR', 'constraint' => '15'),
@@ -201,7 +206,10 @@ class Deployment_hooks_setup_model {
 	 */
 	public function drop_dh_table()
 	{
-		$this->_EE->dbforge->drop_table('deployment_hook_posts');
+		if ($this->_EE->db->table_exists('deployment_hook_posts'))
+		{
+			$this->_EE->dbforge->drop_table('deployment_hook_posts');
+		}
 	}
 	// End function drop_dh_table()
